requirements.txt
llama-index-core
llama-index-llms-google-genai
llama-index-embeddings-google-genai
llama-parse>=0.5.0
python-dotenv